requirements.txt
pandas-gbq==0.26.1
pandas
sqlalchemy
psycopg2-binary
python-dotenv
chardet
kagglehub